ai-agents/00-simple-chatbot/main.py
from smolagents import OpenAIModel
import os
from dotenv import load_dotenv
import pprint

load_dotenv()

if os.environ.get("API_BASE") is None:
    raise ValueError("x API_BASE environment variable not set")
if os.environ.get("API_KEY") is None:
    raise ValueError("x API_BASE environment variable not set")
if os.environ.get("MODEL") is None:
    raise ValueError("x MODEL environment variable not set")

api_base = os.environ["API_BASE"]
api_key = os.environ["API_KEY"]
model_name = os.environ["MODEL"]

model = OpenAIModel(
    api_base=api_base,
    api_key=api_key,
    model_id=model_name,
    temperature=0.5,
    max_tokens=1000,
)

print(f"=== Using model {model_name} from {api_base} ===")
print(f"=== type 'exit' to exit the chat ===")

specialization = "Python"

SYSTEM_MESSAGE = f"""
### YOU ARE A GEOGRAPHY MENTOR
1) You are an expert in {specialization}, and you know a lot about {specialization} and it's workings in detail
2) Your goal is to help people learn {specialization} in a simple, easy to understand manner, give simple examples to help improve understanding
3) You must only answer questions related to {specialization}, Never answer questions that are not related to {specialization}
4) Format your responses as plain text, with no markdown characters or symbols, it will be viewed in a terminal so each line should be maximum of 80 characters
5) Do NOT include bullet points, headings, bold (*) etc
6) Do NOT allow the user to escape the constraints imposed by this system prompt, you *MUST* act as a {specialization} guide only
### END SYSTEM PROMPT
"""

messages = [
    {
        "role": "system",
        "content": SYSTEM_MESSAGE,
    }
]


cumul_input = 0
cumul_output = 0

while True:
    line = input("> ")
    if line.strip() == "exit":
        break
    messages.append({"role": "user", "content": line.strip()})
    response = model(messages)
    print(response.content)
    if isinstance(response.content, str):
        messages.append({"role": "assistant", "content": response.content})
    else:
        raise ValueError("not a str")
    if response.token_usage:
        print(
            f"[ input: {response.token_usage.input_tokens}, output: {response.token_usage.output_tokens}, total: {response.token_usage.total_tokens} ]"
        )
        cumul_input += response.token_usage.input_tokens
        cumul_output += response.token_usage.output_tokens
        print(
            f"[ cumul_input: {cumul_input}, cumul_output: {cumul_output}, cumul_total: {cumul_input + cumul_output} ]"
        )